scripts/BarcodeBERT/requirements.txt
torch
numpy
pandas
scipy
matplotlib
umap-learn
cython
torchtext
transformers
seaborn
scikit-learn
boto3
phate
opt_einsum

